+++ CHANGES.rst
@@ -2,6 +2,13 @@
 CHANGES
 =======
 
+v0.2.0
+-------
+
+* adding refresh token support
+* automatically tracking the expiration of access token and refreshing if necessary
+
+
 v0.1.0
 -------